+++ RELEASE-NOTES.md
@@ -5,6 +5,7 @@
 * Added constructor which takes an `HttpClient` without a base URI (thanks to TimothyMakkison!).
 * Added fluent `request.WithFilter` and `WithoutFilter` methods (thanks to Jericho!).
 * Fixed `resource` parameters for methods like `GetAsync` marked non-nullable.
+* Internal optimizations.
 
 ## 4.1.1
 Released 22 June 2022.